pyproject.toml
[project]
name = "restaurant-bot"
version = "0.1.0"
description = "A restaurant bot with agent handoffs for menu, order, and reservation handling."
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
